requirements.txt
PyYAML~=6.0.2
numpy~=2.1.1
PyQt5~=5.15.11
pydub~=0.25.1